moved yule log service to api service
