--- santa/AngularClient/santa-pone-central/src/app/services/yule-log.service.ts
@@ -1,12 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class YuleLogService {
-
-  constructor(private http: HttpClient) { }
-  private extractData(res: Response) {
-    const body = res;
-    return body || { };
-  }
